signal: unblock async signals before calling instremu [fixes #477]
instremu needs them to interrupt the long runs.
We can safely unblock them even in presence of sas wa, as
the threat is only that the signal can come after the stack
switch but before disabling sas. But for instremu we do not
do the stack switch, we only do it for dpmi_fault().