mapping: use MAP_FIXED when mapping to 0 [fixes #448]
For MAPPING_NOOVERLAP we used to remove MAP_FIXED to see if
kernel can use the address we provide. But this doesn't work
for 0, as NULL is never treated as a hint - kernel takes as
a hint only the addresses above max(PAGE_SIZE, mmap_min_addr).
As a result, use mapping_find_hole() technique.