signal: set up sas via sigaltstack() [#506]
... instead of via direct manipulations with uc_stack.
This only adds one syscall on DPMI client init and one on deinit.
The problem with uc_stack is that it may not be portable.
Plus we don't need to change it on every signal return
(like it was in deinit_handler()).
This allows to stop passing uc_stack around.