signal: stop intercepting SIGTERM [#509]
The cleanup of debugger pipes is now done by dosemu script.
There is nothing that should be _really_ cleaned up, and
SIGTERM should not block (on mhp_exit_intercept() or on some bug).
I find it very bad when an app forces the user to ever apply to SIGKILL.

Also do faster shutdown on SIGHUP.