KVM: fix int 3 and int 4 handling: fixes bp in dosdebug.
DPL is set to 0 in most IDT entries so that software interrupts in DPMI
GPF (which matches Linux kernel handling in native mode). However int3
and int4 (and only those) must have DPL 3 as in the Linux kernel so
they generate traps 3 and 4 instead of GPFs for INT3 and INTO instructions.