fatfs.c: check file size of FreeDOS kernel to load
As the boot sector with BPB is located at 1FE0h:7C00h,
the kernel must be smaller than 1FE00h+7C00h-600h
= 27400h (ca. 156 KiB), to avoid overwriting the BPB.
(The BPB can be located elsewhere, and its actual
address is passed to the kernel in ss:bp, but we use
the default address as implemented by their loaders.)

This commit makes the handler check for the load address
plus size being below 1FE0h:7C00h - 8192, where the -8192
is a stack reservation to insure the stack doesn't overwrite
the end of the kernel. A file larger than this is rejected.

This is equivalent to the check I added to grub2's FreeDOS
direct loading support, which implements the FreeDOS load
protocol too.
Refer to http://git.savannah.gnu.org/cgit/grub.git/commit/?id=08bcec502084e322318ecad1bc4c04d3ddc10d32

The original FreeDOS boot sectors do not check the size of
the kernel they're loading, but this must be so only because
there is no space in a boot sector to check for that. Refer
to https://github.com/PerditionC/fdkernel/tree/master/boot