dosdebug: Breakpoint DOS code may be overwritten [fixes #591]
Previously when the DOS code referenced in a breakpoint was overwritten
by DOS, the breakpoint address was silently zeroed but the validity
field was not changed. Typically this was seen in the example where the
bootloader was being debugged. Setting a BP at 0070:0000 as the first
dosdebug command after starting with -H1 then expecting 'g' to stop at
it was generally unsuccessful.

1/ Correct the current behaviour
  a) Add a message to report that the code was overwritten
  b) Set the is_valid field to false so that it no longer appears in
     breakpoint listing 'bl'

2/ Add new debugging bit to allow the debugger to track the code change,
  save the new opcode and reinstall the INT3 opcode. To use this function
  to debug the bootloader combine 1 & 8 to start dosemu with -H9 option.