bios.S: Avoid copying over text font [fixes #552]
The lower 128 characters of text font are kept in the bios at f000:fa6e
and we copy the bios in two chunks to seemingly avoid overwriting this.
However at present the bios_f000_endpart1 and bios_f000_part2 symbols are
pointing to the same address so rendering the two part copy ineffective.

>>> Existing bios.o objdump
0000f5d8 g       .text	00000000 bios_f000_endpart1
0000f5d8 g       .text	00000000 bios_f000_part2

This patch defines the bios_text_font symbol, positions it at the 0xfa6e
offset and advances the origin to avoid overwriting the text font that
may already be defined in the bios.

>>> New bios.o objdump
0000f5d8 g       .text	00000000 bios_f000_endpart1
0000fa6e g       .text	00000000 bios_text_font
0000fe6e g       .text	00000000 bios_f000_part2