dosdebug: Rename symbol functions to reflect use
The getaddr, getaddr2 and getaddr2u function naming is confusing so
rename them to reflect their use. The getname2u name seems very odd,
but it most definitely operates on the second symbol table which is
used for DOS.
  getname   -> getsym_from_unix
  getname2  -> getsym_from_dos_segofs
  getname2u -> getsym_from_dos_linear

Rename the opposite functions for consistency
  getaddr  -> getaddr_from_unix_sym
  lookup   -> getaddr_from_dos_sym

No other changes are made.