commands: change MaxCmd to AmountCmd, use jb conditional
The first change is to rename the MaxCmd label to AmountCmd, because the
current way that this label is created actually calculates the *amount*
of table entries (ie commands), not the maximum command (one less).

The second change is to use an unsigned comparison conditional branch,
ie "jbe" instead of "jle". With "jle", values that in two's complement are
below zero would incorrectly try to dispatch through the table.

The third change is to use "jb" instead of "jbe", which relates to the
different sense of the calculation as already indicated by the first change.
A command index needs to be *below* the AmountCmd value, not below-or-equal.