dosdebug: Allow trace to stop on the right address
This commit allows the 't' trace command to follow the expected program
flow when encountering an interrupt call. For example;

0fc2:0113 B409             mov  ah,09
0fc2:0115 BA4E01           mov  dx,014E
0fc2:0118 CD21             int  21
0fc2:011a 90               nop

0fc2:011b CD23             int  23

With this patch after 'int 21' is 't' traced, the debugger stops and
waits for execution of the 'nop' instruction. Depending upon what
instruction was used in the interrupt handler to return, the previous
behaviour was:
  a) IRET - the debugger stopped one instruction too late i.e. 'int 23'
  b) RETF 2 - the debugger resumed running.

Fixes #468