mkpluginhooks: stop checking for configure
This fixes the recent build regression that is after
"make realclean && make" you don't get any plugins working.
"make realclean" removes configure, and mkpluginhooks runs
before it is re-created by make.

This is actually quite a problem, because it means that
configure can be removed but the plugin's enable info, which
initially depends on configure - stays valid. Of course
"make realclean" should remove the plugin enable info, but
currently it does not. Instead this patch marks configure
as .SECONDARY target, meaning that we don't need it after
Makefile.conf is created. "make clean" is adjusted to remove
configure.
So now the rule is: configure is only needed to generate
Makefile.conf and config/plugin_enable, after which it can
be deleted.