MFS: Display disabled drive state in lredir
If Dosemu has a drive configured for redirection, but the CDS flags are
not set for DOS to recognise it, have lredir display the DISABLED
attribute.

Note: This patch removes the CDROM unit number display code from lredir,
as it was ineffective because the corresponding bits were never set in
mfs:GetRedirection().

[Fixes #695]