MFS: Some DOSes may move SDA and LOL in memory
Starting with EDR-DOS 7.01.07 the SDA and LOL memory location may move
after MFS has been initialised. This seems to be an effect of a new
feature to allow the user to choose where to place certain parts of
DOS (LOW, HIGH or UMB).

EDR-DOS 7.01.07 and EDR-DOS 7.01.08 WIP (21.7.2011) no longer crash as
reported in issue #588