bios: Don't link BIOS code directly into Dosemu
As a precursor to making Dosemu more platform agnostic, don't link
the BIOS directly into Dosemu. Assemble the BIOS code into a flat
binary file and include the result as data instead. This will
allow a different toolset to be used to create the BIOS if desired,
for example i386-elf-*. Issues #732 #721