LFN: Fall through if the drive isn't ours
In the case that we don't own the drive in question, for example it's
FAT, we should fall through to a further LFN handler (even though we
may be the last) instead of authoritively returning a drive error to
the caller. The caller may choose to see this lack of LFN support as
an indication to fallback to SFNs as required.

[fixes #770] [#539]