coopth: add coopth_abandon() function
Mostly same as coopth_leave(), but to be used from leavedos().
This allows to simplify coopth_leave() that should not deal with
leaving the detached threads. The only point to "leave" the detached
thread, is to run its post handler from the appropriate context.
Otherwise leaving the detached thread is completely pointless, so
the assert is added to coopth_leave().
Also coopth_abandon() silently discards the non-permanent post handlers,
and coopth_leave() now errors on them.