Revert "MFS: Display disabled drive state in lredir"
This reverts commit 9a0a9643a2260dfb5882fca25f6c2a4e2e14ea02.

This patch is wrong.
It assumes that cdrom flags are not in effect, but they
actually come from the "read_only" integer field.

I would suggest to not hijack CX for device flags, as it
is documented to be the user-supplied value. We can use
DX for our extensions instead.

Reopens #695