dpmi: make realmode stacks combined for all clients [#898]
We used to have the set of realmode stacks (6) per client.
Client couldn't use the rm stack of another client.
With this patch the rm stacks form the single "pool" to
which every new client adds a set of stacks.

32rtm jumps to the return address and stays resident, and
if it was started from some dpmi client (eg comcom32), then
that client is no longer current. When it returns to PM, the
rm stack counter of 32rtm goes below zero.
This patch removes the per-client rm stack counter hence
avoiding the (part of the) problem.