ne2000: Need to serialize the interrupts
If a packet is received before the post send packet interrupt is
triggered, the pic_request() is generated based upon the state at
transmit time. To avoid this we just return from
ne2000_receive_req_async() without reading the incoming packet if
we marked ourselves busy at transmit time, we can process it the
next time we are called.