msdos_ldt: fix alias management regression [fixes #926]
Needs to pre-fill back-buffer with an actual ldt content.
It could do w/o pre-filling in the past (even though it actually
did the pre-fill) because the alias management was initialized
much earlier (at dpmi_setup() stage) and there was nothing to pre-fill.
Now it is done at client init stage.