dpmi: implement NX support [fixes #936]
This is a DPMI extension.
Used some "must-be-zero" bits in DPMI structures for the NX flag.
We do not need the shared memory for LDT buffer to be executable,
so use the NX bit to explicitly specify that.
Cases where we do need the executable shared mapping, fortunately
all go via mapping drivers, which would in this case use anon-shm.