mfs: maintain list of handles per sft [fixes #938]
This fixes the recent regression of 3be5394.
The problem was that multiple opens for the same file
lead to the creation of multiple handles. Some of them
can later be closed and re-used. But sft is not invalidated
untill _all_ its handles are closed.
And so duplicate handles started to exist, leading to a
wrong file look-up.