debug: get rid of dpmi_mhp_TF
It was needed to hide the TF flag from client.
But it severely limits the functionality, for example you
can't have the hw interrupts untraced while tracing only sw
code path, at least its not simple to do.
Better to use normal TF techniques and not to care about
the anti-debugger tricks.