dpmi: exit control thread before switch to RM [fixes #979]
It appears that sometimes the DPMI thread had no chance
to exit before the switch to RM happens. Even worse, another
thread can be created when switching back to PM, and the
first thread is still there too. In that case the first
thread will not exit as it will see the the PM active.
So there will be 2 DPMI threads running concurrently!
The recent sanity checks started to trigger that bug.
I think this fix will improve the overall stability, but
that remains to be seen.