Extend direct IDT use to software ints 0x11 to 0x20.
This is possible since those never trigger as exceptions in our CR0/CR4 setup.
int 0x10 is still special since it can either be a floating point exception
or video software interrupt. In real DOS this is managed by not enabling
cr0.ne so FP exceptions trigger IRQ13 directly but KVM forces this to be set.