fatfs: Move RxDOS.2 loadtop calculation further up into a variable
fatfs: Implement RxDOS.3 load protocol size limit

This allows files that are larger than the amount of available
Low Memory Area or, in our case, larger than the space between
the load address 00700h and stack/LSV/BPB at 1FE0h:7C00h - 8192.
This is allowed because the lDOS-based load protocol allows
loading a partial file, if at least 1536 bytes are loaded. The
iniload stub can load the remainder, if needed, by itself.

This feature is intended primarily for kernel files that carry
additional data, such as a .ZIP archive, behind the actual
kernel and .EXE mode executable images. Loading the entire file
would fail, as it is easily too large, but loading as much as
possible then handing over control succeeds, because the iniload
knows how much of the file is actually needed for execution.

fatfs: Finish RxDOS.3 load protocol implementation

Aside the added note about the top 20 KiB, this changes the
load segment to 200h (8 KiB) which is preferred now. This
insures that no sector read crosses a 64 KiB boundary, even
with the theoretical maximum sector size of 8 KiB.

The changed load segment is compatible to all RxDOS versions
(as well as the bootable lDebug based on lDOS boot iniload).
Free load segment choice was always part of the protocol.

A part of the RxDOS.3 load protocol that this does not add
is the check for the 2-byte string "lD" at offset 1020.
This check would require seeking, reading the file, then
seeking back in the current mimic_boot_blk implementation.