fatfs: Implement RxDOS.3 load protocol size limit
This allows files that are larger than the amount of available
Low Memory Area or, in our case, larger than the space between
the load address 00700h and stack/LSV/BPB at 1FE0h:7C00h - 8192.
This is allowed because the lDOS-based load protocol allows
loading a partial file, if at least 1536 bytes are loaded. The
iniload stub can load the remainder, if needed, by itself.

This feature is intended primarily for kernel files that carry
additional data, such as a .ZIP archive, behind the actual
kernel and .EXE mode executable images. Loading the entire file
would fail, as it is easily too large, but loading as much as
possible then handing over control succeeds, because the iniload
knows how much of the file is actually needed for execution.