fatfs: Finish RxDOS.3 load protocol implementation
Aside the added note about the top 20 KiB, this changes the
load segment to 200h (8 KiB) which is preferred now. This
insures that no sector read crosses a 64 KiB boundary, even
with the theoretical maximum sector size of 8 KiB.

The changed load segment is compatible to all RxDOS versions
(as well as the bootable lDebug based on lDOS boot iniload).
Free load segment choice was always part of the protocol.

A part of the RxDOS.3 load protocol that this does not add
is the check for the 2-byte string "lD" at offset 1020.
This check would require seeking, reading the file, then
seeking back in the current mimic_boot_blk implementation.