DPMI native: don't change fpregs pointer in sigcontext
At least on i386 kernel 4.19 is picky about the size of the area that fpregs
points to. We know for sure that what fpregs points to on signal entry is
fine so don't change and copy FPU state to it instead of setting the pointer.

Without this fix I got hard SIGSEGVs (no backtrace) on i386 native DPMI as
soon as a DPMI prog is started and dmesg messages such as
"dosemu.bin[32058] bad frame in rt_sigreturn frame:30e78129 ip:17a1 sp:7f0
 orax:ffffffff in memfd:dosemu_32058 (deleted)[0+b8000]"