Use $CC -dumpmachine to determine default target machine.
This was useful for me as I have a mostly 32-bit userspace but often
run with a 64-bit kernel, and the default target for gcc is then -m32.

The compilation should then choose between -no-pie and -pie depending on
GCC's (or Clang's) default target instead of the kernel's arch `uname -m`
that is running during compilation.