vga_emu_fault: Pass cr2 as dosaddr_t & error code directly
This allows it to work with the 0-based separate vm86 memory mapping, and
allows some cleanups in the KVM code as it no longer needs to setup a
sigcontext for only _cr2 and _err. scp is now only passed to vga_emu_fault
for DPMI, so the pmode parameter is obsolete.