vm86: create extra lowmem alias for base 0
The creates an extra lowmem alias at base 0 for the exclusive use of
vm86(). Memory from 0 to 0x110000 is not used by DPMI nor by DOSEMU itself
which always uses mem_base'ed relative memory (so that can be contiguous
now it no longer needs to be zero-based).

This means double mappings for low memory for alias and kmem maps, and
adjustments on the side for mprotect and munmap.

This allows PIE on i386 as the address space is more flexible this way.

Fixes #620.