cpu-emu: fix page fault handling in simulator.
This fixes a regression caused by b3ee7dfb0 which eliminated the global P0.
However the sigsetjmp cannot be in Gen, since Gen calls don't quit after
a page fault, so it needs to be higher in the call stack, _Interp86 is
a good place since P0 is available to retrieve eip; it just needs to be
volatile.