dpmi: look for common pattern for cli w/a [fixes #1031]
Make sure the sequence is:
pushfd
pop eax
cli

This seems to be a common pattern for DOOM problem, and
the epilog is:
push eax
popfd

But we can't check for an epilog, as its distance varies
a lot.

This patch makes sure other programs do not hit the w/a
and get their cli ignored. Transport Tycoon seems stable
now even with the weakened interrupt re-entrancy protection
of 89f7a8f.