simx86: mostly revert acee39e (extensive use of Cpatch)
Cpatch can't be used for DPMI page faults (e.g. with DJGPP null pointer
protection as tested in test-i386.exe: that test is fixed with this change),
so we must check if the address is either VGA memory or protected by cpuemu.

Using Cpatch is a strategic place in e_vgaemu_fault allows using Cpatch
on VGA memory then but unfortunately that still produces glitches in Jazz
Jackrabbit (which are also there without the revert).

The one thing left from acee39e is to allow Cpatch for code hits.