simx86: replace use of O_MOVS_MovD without REP with O_MOVS_LodD/StoD
The use of LODS/STOS is temporary, they will be replaced with L_DI_R1
and S_DI later. This way there is no more MOVS in generated code without
REP, so the cpatch and sigsegv code can also be simplified. See also #1033
(not fixed completely for Jazz yet) and #156.