simx86: fix floating point flagging of nodes in JIT
test-i386 was randomly failing floating point tests where the FP environment
was not properly set restored. The cause was that for code sequences that use
FP that FP flag was not always set, which was the fault of the node linker.

The node linker needs to not only flag FP for the node that links to the new
sequence but also any other nodes that link to it, so I added a recursive
function to flag all back references. It can stop at any back reference that
was already flagged since those were treated earlier with all their back
references.

test-i386 now passes completely for "full" JIT cpuemu with FreeDOS so I
added it to the Travis tests. For FDPP it still fails with a GPF for unrelated
reasons, so skipping it there for now.