Proof-of-concept faultless JIT
This basically does:
* apply Cpatch unconditionally for every write
* no longer mprotect DPMI memory
* still mprotect low memory for compatibility with JIT+native, but JIT
  writes via alias to low memory
* add a counter

this is for testing only, not to be merged. It will fail the Travis tests
but only because of unhandled DPMI page faults.

In my testing it still faults on PUSHA which causes a few faults when you
running Win31.

I'm just curious how this does in your vm86jit benchmark @stsp.