simx86: eliminate TheCPU.veflags for vm86
Just keep those flags (ID,AC,NT, which in v86 mode just serve CPU detection
algorithms) in TheCPU.eflags, as there is no danger having them
there unlike in real flags. IOPL is a bit special but pushf will need to
set that to 3 on the stack image anyway so a program can't flip it visibly.

This makes pushf, popf, int and iret a little simpler so easier to compile.