simx86: fix int_revectored regression from 6f16dbf3 [fixes #1181]
btl encoding with immediate value is a bit tricky.
The value must be split to high and low order bits.
See http://x86.dapsen.com/html/file_module_x86_id_22.html:
"Some assemblers support immediate bit offsets larger than 31 by using
the immediate bit offset field in combination with the displacement
field of the memory operand. In this case, the loworder 3 or 5 bits
(3 for 16-bit operands, 5 for 32-bit operands) of the immediate bit
offset are stored in the immediate bit offset field, and the high-order
bits are shifted and combined with the byte displacement in the addressing
mode by the assembler. The processor will ignore the high order bits if
they are not zero."

And it indeed does ignore high order bits!
So lets follow the recommended split.