simx86: fix InvalidateNodePage() [#86]
Tree_InvalidateNodePage() was mistakenly using higher bound instead
of lower when searching "less-than" node, thus sometimes finding
"greater-than" instead.

That caused stalled code nodes w/o code marks.
Most importantly that was causing the codemark "corruptions",
i.e. the codemarks could be set for the beginning of the node,
but when node crosses the page boundary - no more code marks
for it. So it was not caught by the recently added e_querymark()
check.
But was caught by the e_querymark_all() code from the previous patch.

With this patch the opposite is possible: you can now have stalled
code marks if the node was crossing the page boundaries. But I
don't think it is a big problem. Can be fixed later.

This fixes most of the bug #86.