simx86: add cache for FindTree
By adding a hash-table like cache to FindTree, indexed by the low 12 bits
of the PC we can obtain a 99.99% hit rate and speed up searches so FindTree
is sped up by a factor 3, and goes from 12% to 4% of total execution time.

This obsoletes most uses of LastXNode.

See also the discussion at #86