emm: check handle_info.active, not .numpages > 0 [fixes #1192]
Refer to the bug report for details. A handle that is active
(allocated) but has zero pages allocated to it is valid and
should appear in the table returned by function 4Dh. Otherwise,
the table would carry uninitialised entries at the end for
each of the zero-page handles.

The second change here is to return (in bx) the value of the
variable tot_handles. This is the amount of entries actually
entered into the table. While tot_handles and handle_total
should always be the same, this is more robust in case they
are not. (In the case of the bug fixed by this, the final
table size returned would have been more correct if the code
had set bx to tot_handles already.)

Note that function 4Bh returns handle_total because it does
not enumerate the active handles itself. This function is
used by applications (such as lDebug) to determine how large
their buffer has to be to call function 4Dh. So a smaller
handle_total than tot_handles would be a severe error.

Therefore, we catch cases where tot_handles would grow beyond
handle_total and cancel the loop at that point. An error
message is then displayed to inform the user. Another error
message is only displayed after the loop, if at that point
the two variables do not match. (In practice this second
error only occurs when tot_handles is below handle_total,
which is unexpected but not a buffer overflow error.)