simx86: check every gen addr to avoid overlaps [fixes #86]
(with help from Bart)

When generating code, check every PC address to avoid code
nodes overlap.
This should likely slow things down a lot, and there is a lot to
optimize in that patch series, but lets start from the slow but
correct code, rather than from fast and broken.

This was the most difficult bug in the entire dosemu2 project's
life span. Given that it didn't require any reverse-engineering
efforts on DOS side, and yet was times more difficult than the
ones that do require reverse-engineering, its a reason to celebrate. :)