Revert "cpuemu: extend protected part to int xx instructions."
This reverts commit c1ddb275b8ca54fe66b8b6144cf0bb5c861d8f76.

Conflicts:
	src/base/emu-i386/simx86/codegen-x86.c

Closes #1209

This was a speed-up trick needed because the INT node won't be
linked with the code before it. And the INT was not compiled in
back then.
So the hack was to make sure that when/if INT is patched to
something else (by an FPU emulator, see #982) that CAN be compiled,
then the previous node is destroyed too, to have things re-jitted.

Now however INT is compiled in v86, and is TODO for a prot mode.
So the hack now only relevant for prot mode.

Revert it as it may cause troubles for further developments.