simx86: handle mov ss/pop ss correctly for compiled code (see #1147)
CeS_LOCK was only set in the interpreter and not set correctly when executing
JITted code. To handle this introduce a new TNode flag F_INHI
(INHI=Inhibit Interrupts) and also rename CeS_LOCK to CeS_INHI to remove
confusion with the LOCK instruction.

The tests in src/tests test this in protected mode where it still goes via
the interpreter. I tested it in vm86 mode too in DEBUG single stepping through
pop ss and seeing that that would skip the instruction after pop ss.