simx86: fix two contiguous inhibiting instructions [#1214]
if we have two "mov ss" or "pop ss" instructions in a row, the second one
should not inhibit interrupts again. For the interpreter this was most
easily done using a new CeS_MOVSS flag, which can then be checked in
combination with CeS_INHI, which is then set for the first instruction,
and checked and reset for the second.

The logic now handles CeS_TRAP/CeS_SIGPEND itself so HandleEmuSignals no
longer needs to fiddle with CeS_INHI.