dautoemu.bat: dont set PATH to dosemu tools
This works around the novell command.com bug, that is
described in RBIL as follows:
---
BUG:    Novell DOS 7.0's COMMAND.COM (prior to Update 12) will attempt to run
          a disk program with the indicated name even if the returned length
          byte is zero, because the register used to flag this case is
          clobbered without first checking it.
---

As the result, dosemu commands are run twice.
But if we remove them from PATH, then they will run only once,
followed by a "bad command or filename" message. Which is much
better than to run twice.