Travis: Disable lfn file info test on FreeDOS 1.20
Since the workaround to avoid FreeDOS <= 1.20 int2f/1217 bug was removed
in 43e66026 lredir of new drives no longer works on those versions. For
this test we need to be able to create a sparse file of up to 6 GiB on
an MFS drive, but this can't be in a directory that would first be used
to create a fatfs drive (for 2 reasons: fatfs is FAT16B; file gets filled
out to consume space), so let's add it as a known failure.