simx86: call InvalidateSegs() from emu_update_LDT()
simx86 caches segment bases, which need to be invalidated if the LDT changes.
Up so far the invalidation happened unconditionally for "int 31" but that's
sometimes not necessary and not always sufficient, for instance emulated direct
LDT writes, and int21/ah=4c to terminate the DPMI client. I triggered the last
issue when using KVM vm86 + emulated DPMI.