build: update configure by deps
plugctl.sh was not updating configure if configure.ac changed!
It was only building it when it doesn't exist, but not when
outdated.
As the result, the configure.ac changes could get "unnoticed",
resulting in a subtle build breakages, or "fdpp unavailable" error.
Quite a horrible bug that seem to have beaten a few ppl on the net.