Revert "dosdebug: disable interrupts while tracing" [#1292]
This reverts commit b4dc178dc27e79701f3e947bf1311d33f35f2aa8.

Tracing may alter IF, so we either need to sync it or revert.
But we can also stop by breakpoint, not only with "stop" cmd,
so there needs to be many sync points.
But the worst is that during trace we will see IF cleared,
or we also need to set if before step trace, then clear, etc.
Much easier to just revert.

Conflicts:
	src/tools/debugger/mhpdbgc.c