allow use of rdtsc on machines with tsc_reliable
dosemu traditionally disables the use of rdtsc on machines with more
than one processor, this is because older machines would not keep the
tsc synchronized across cores.

There were also machines that stepped down to save power but not
increment the tsc at the same rate, confusing anything using it. This
has not been a problem for several hardware generations, and modern
machines will get the reliable_tsc flag if they can be used for timing.

This patch checks for that flag, and then allows rdtsc on SMP machines,
which can be a big performance win.

Additionally, some logic was removed that supported old kernels before
2.1.74. The code wont break on a kernel that old - just fall back to CPU_386.

The code used to check of the cpuinfo contained a "features" line, but I
checked all the way back to 2.0.39 and it was still called "flags", so
that must predate the 2.x kernel series (1996 or so).