int: simplify drive redirection management
There is no need to do a add/lookup round-trip to mfs for
redirected drives to find their number. It became known
after redirection, so we can return handle that will help
to obtain the needed number later, without any extra look-up.

This spares CX register because user-data is then not needed.