dpmi: bind PM client num to return address from rm [#1336]
Binding to RM stack doesn't work because the client can provide
its own RM stack. In that case its difficult to find the stored
value in it, as you need to first remove the ECX number of arguments.
But you need ECX of the client whose number is not yet known.
The approach used in 6488d28b works only when the stack length
is known, which is not the case when client provides it.

So lets just allocate 32 return points (one per client) and bind
the client number to that.