simx86: fix SINGLESTEP hang of dpmi jit [#1380]
In SINGLESTEP mode the nodes were removed from avl w/o invalidation.
This means the pages were left protected, and querymarks - orphaned.
This resulted in a complete hang of entire dosemu, as the signal
handler was calling InvalidateNodeRange() but it was not doing
anothing at all as the nodes were removed.