msdos: more work on THUNK_16_32x extension [fixes #1341]
Make it controlable via BX and make it per-client.
RSP should disable it before calling to 16bit code and re-enable
after, so it is not a problem to have RSP to also enable it per-client.

This is a difference with THUNK_16_32 extension of DPMI: one can't
enable it per-client because it is needed for the per-client
call-back to work. But THUNK_16_32x is only used for debug prints.