dpmi: implement LDT_MONITOR extension and API
We can't call the monitor directly from LDT update functions
because they will return status to the caller later, which will
corrupt our call state.
So I accumulate the LDT updates in a bitmap and process it
before the next entry to client code.

Convert msdos_ldt to that new API.