conf: add $_vga_fonts option
It selects whether to render native or VGA fonts.
This allows to set up the native font configs to some sensible
value by default, rather than to "" as we did before.

Note: this makes no sense as long as we do not support unicode
fonts. This is for the future unicode font support.