conf: implement $_file_lock_limit [#428]
This is needed to work around the bug in the programs that
probe the amount of supported locks by applying them infinitely
in a loop.

Note: single unlock resets the counter because, with lock coalescing,
we don't really know how many locks remained.
This is just to work around bugs, not to enforce any limits.