mouse: call user handler on separate stack [fixes #1461]
User handler should always be called on a separate stack,
or the stack overflow can happen.
In this particular case we switch away from the DOS irq
stack, which is quite wrong by itself.
Either disable the DOS irq stacks completely, or revector
the mouse interrupt the way the DOS irq stack is avoided.