dpmi: revert 3cc5ef0ecb [fixes #1457]
Realmode stacks cannot be "flat", or switching between PM
and DOS prompt doesn't work in win31.
When returning after switch, the RM stack of the switched
out client should be preserved. But after 3cc5ef0ecb it
because "unwound" even after the switch to another client.
It was then re-used and corrupted, so the switch back was
not possible.

Note that 3cc5ef0ecb was obsoleted by 8330d87ee.
The initial problem was that 32rtm's stack counter can became
negative, but after 8330d87ee we would return to comcom32 and
correctly decrement its stack counter. So 3cc5ef0ecb had to be
reverted much earlier...