Test: Be mostly specific about what testcases get run
Now that Travis repository is fixed the tests run again on the cron
job, but now we are hitting the maximum time that Travis allows for
a single job. We need to reduce what's tested on the cron job, but
the skipping mechanism wasn't that great, so just specify the
testcases directly.