Revert "kvm: check for KVM_CAP_SYNC_MMU extension [#1492]"
This reverts commit 53ef318a7258d7f84d59735a83d54dd9635b59f1.

KVM_CAP_SYNC_MMU description in kernel docs seems invalid.
It says:
---
When the KVM_CAP_SYNC_MMU capability is available, changes in the backing of
the memory region are automatically reflected into the guest.
---

But it seems like it is actually only needed when we mmap()
that memory region to something else.