SDL: Create GL contexts for the async threads.
This band-aid makes sdl.hwrend work with Nvidia.

It's just a dirty fix though. Dosemu2 is using SDL in a way it should not
be used - renderers are not thread safe:
https://github.com/libsdl-org/SDL/issues/986

A better approach would be to create the SDL objects from the render
thread and only use them from there. That requires some restructuring
of the entire file; I tried 3 times and ran into dead ends. Maybe it needs
a restructuring of the larger video output system.

It also appears that dosemu is doing pixel format conversions (e.g.
palettized to 32 bit RGB) before calling into sdl.c. Ideally, we'd create
an SDL texture that matches the VGA mode (e.g. GL_RED8 for 8 bit palettes),
and do the conversion in a shader with a second texture for color lookups.

Beyond that, we should enable vsync and use it as a rate limit in the
renderer thread. Hypothetically we could then feed the return timings
from SDL_RenderPresent() to the application via VGA interrupts.