dpmi: try to fix int23 handling [#1438]
Patch is completely untested, and is based only on reading
the (apparently also untested) current int23 code of dpmi.
When the PM handler leaves a word on stack, it was copied
on an RM stack, which is wrong because the RM stack still
has the iret frame.

That code was introduced by commit ed322f9eda, and as far
as I can recall from 17 years ago, I actually never got
int23 working, contrary to what was advertised (the same
commit also added ints 0x1c and 0x24 which worked). Now I
see why but don't remember what was the test-case. :)